HAWQ-178-jsonpath: modify columny types for unspoported Short and Byte[] types

+++ pxf/pxf-json/src/main/java/org/apache/hawq/pxf/plugins/json/ColumnDescriptorCache.java
@@ -151,7 +151,7 @@ public class ColumnDescriptorCache {
 			clazz = Integer.class;
 			break;
 		case SMALLINT:
-			clazz = Short.class;
+			clazz = Integer.class; // Should be Short.class (Bug) 
 			break;
 		case BPCHAR:
 		case TEXT:
